+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Pvp fetcher fetching when user doesn't have pvp permissions
 - Guilds of the day to be cached for the day
 
+### Changed
+- Fetch logging to be more granular
+
 ### Added
 - Api token invalidator